LayoutTests/http/tests/security/contentSecurityPolicy/block-favicon-expected.txt

CONSOLE MESSAGE: TypeError: testRunner.setIconDatabaseEnabled is not a function. (In 'testRunner.setIconDatabaseEnabled(true)', 'testRunner.setIconDatabaseEnabled' is undefined)

http://127.0.0.1:8000/security/contentSecurityPolicy/block-favicon.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/security/contentSecurityPolicy/block-favicon.html, main document URL http://127.0.0.1:8000/security/contentSecurityPolicy/block-favicon.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/security/contentSecurityPolicy/block-favicon.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/security/contentSecurityPolicy/block-favicon.html, http status code 200>
http://127.0.0.1:8000/security/contentSecurityPolicy/block-favicon.html - didFinishLoading
